Modifications to the app.
I've added functionality that the nodes are informed to update something
instead of sending the result. This causes some synchronization issues,
but should be solved with a lock {} in the Update method. This is used
because it could inform the nodes to make an update, and after that it
will release the token or something similar. This could cause for it to
make another update (make another operation) instead of waiting for it
to finish the first one. Could have different results in the end.